add teacher/student class and a teacher specific module
