+++ student.rb
@@ -1,15 +1,10 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'teacher_class'
+
+class Student < TeacherStudent
 
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
